extract TrackList into separate component
