+++ app/(client)/tracks/top/page.tsx
@@ -1,8 +1,8 @@
 import { getTracksAudioFeatures, getUsersTopItems } from '@/lib/spotify';
-import Image from 'next/image';
-import ExportPlaylistButton from '../components/ExportPlaylistButton';
 import { Suspense } from 'react';
 import Loading from './loading';
+import TrackList from '../components/TrackList';
+import ExportPlaylistButton from '../components/ExportPlaylistButton';
 
 interface Props {
   params: { slug: string };
@@ -36,11 +36,6 @@ export default async function TracksPage({ searchParams }: Props) {
 
   const { items: tracks }: { items: SpotifyApi.TrackObjectFull[] } = response;
 
-  const getArtistString = (artists: SpotifyApi.ArtistObjectSimplified[]) => {
-    const artistNames = artists.map((artist) => artist.name);
-    return artistNames.join(', ');
-  };
-
   const getTrackIds = (tracks: SpotifyApi.TrackObjectFull[]) => {
     return tracks.map((tracks) => tracks.id).join(',');
   };
@@ -68,32 +63,7 @@ export default async function TracksPage({ searchParams }: Props) {
 
       <Suspense key={searchParams.timeRange} fallback={<Loading />}>
         <ExportPlaylistButton headerText={headerText} uris={trackUris} />
-        <ul className="flex flex-col gap-4">
-          {tracks.map((track, index) => (
-            <li key={track.id} className="flex gap-4">
-              <p className="font-bold text-md self-center">{index + 1}</p>
-              <Image
-                src={track.album.images[1].url}
-                alt="a track image"
-                width={50}
-                height={50}
-                className="rounded-xl w-24"
-              />
-              <div className="flex justify-between w-full">
-                <div className="flex-col mr-">
-                  <p className="text-xl font-semibold">{track.name}</p>
-                  <p className="font-extralight">
-                    {getArtistString(track.artists)}
-                  </p>
-                  <p className="font-extralight">
-                    BPM/Tempo: {Math.round(track.tempo)}
-                  </p>
-                </div>
-                <a href={track.uri}>track link</a>
-              </div>
-            </li>
-          ))}
-        </ul>
+        <TrackList tracks={tracks} />
       </Suspense>
     </main>
   );
